Modify loop timing
Modifies the loop timing to achieve the configured rate regardless of loop processing time. If the loop takes longer than UPDATE_INTERVAL_SEC, the delay between loops will always be at least MIN_LOOP_DELAY_SEC. This should probably be set to whatever minimum time between polls the sensors may require, and it needs to be at least 0, so that a really long loop will never try to sleep for a negative amount of time (raises ValueError).

Tested in python3 on my desktop.